Fix a memcpy() that used the wrong size.
When distributing all the nodes' initial memory region tables around the
job, when copying each node's initial table entries from the allgathered
data to our map we were copying enough bytes for the total size of the
entire table, including its header and all the entries allocated for it.
That's way too many; the target has that much room, by definition, but
the source doesn't have that many valid bytes.  We should instead copy
just enough bytes to cover the maximum number of table entries that are
in use on any node at this point.  That's all that was transmitted in
the allgather.