Fix .bad file that I messed up
This was a classic case of filing the future, then making last-minute
changes to make the test "better" and failing to update the line
numbers as a result of the changes.  Sloppy.  Worse, I was doing
post-merge testing on master yesterday to make sure I hadn't done
anything like this, but the valgrind run took so long that I didn't
get to the linux64 + futures case.  Sorry for the noise.