Adapt the task=qthreads shim to CHPL_RT_OVERSUBSCRIBED.
If CHPL_RT_OVERSUBSCRIBED is set at launch time, set QT_AFFINITY=no
and QT_SPINCOUNT=300 in the Qthreads tasking shim.  (It's an open
question what to do if this is set on a Cray X* system.  That's not a
combined configuration we ever use, really.  For now we're ignoring
the 'oversubscribed' setting on that target.)