Add common code for setting environment variables.
This adds chpl_env_set() and chpl_env_set_uint(), which set a given
environment variable to a string or uint64_t value, respectively.
This reduces code size because failure checking can be in the common
functions, standardizes the message printed in the unlikely event an
error occurs, and provides a place for us to record what env vars are
set in launchers so that verbose output can include that info in the
future.