Allow `exit(0)` to work with ugni
Previously, doing an `exit(0)` would hang under ugni on slurm systems.
This is because only locale/process 0 would actually stop executing and
slurm doesn't do anything to kill the other locales/processes. This
updates ugni to make more of an effort to shutdown remote nodes when an
uncoordinated exit is initiated. Specifically, we'll unblock the main
process waiting on the shutdown condvar and stop the polling thread on
each node. This leads to all nodes exiting if there's no outstanding
comm, but if some task is still waiting for comm to complete, we don't
do anything cancel it.

So this improves the simple case of `exit(0)` being called from locale 0
when no remote comm is occurring, but can still hang in more complicated
situations. At the moment, I think that's probably sufficient and don't
think it's worth the time to implement more comprehensive uncoordinated
shutdowns.

Resolves 7521

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>