Enable BlockDist bulk-transfer by default
This commit enables bulk-transfer by default for the Block distribution.
This optimization now supports strided domains and rank-changes.

In order to support this change, a new function has been added to
DSIUtil which translates a subset of a domain to another domain's index
set.  This new function can handle rank changes, and is easier to use
than the old 'bulkCommConvertCoordinate' helper.

The 'computeActiveDims' function from DefaultRectangular has been
renamed and moved into DSIUtil in order to support the
domain-translating function.

The BlockDist bulk-transfer implementation has been improved by only
operating on locales that own indices involved in the transfer. This is
supported by the 'activeTargetLocales' helper method.

The implementation of Block=Block has been simplified to use
domain-translation instead of computing chunks in some iterator. This
simpler implementation results in less communication and supports
strided and rank-changed domains.