Only unload hugepages if found && check chpl-vers at compile-time
The previous buildMason implementation assumed hugepages was loaded,
which caused a build hang for systems without hugepages module. Now we
check for the module explicitly.

Also added a `chpl --version` call to buildMason so that we can provide
the chapel version as a compile-time value as opposed to determining it on
the fly.

Add pre-release info to MasonUtils as well

Revert embedding chapel version