Have Py3CompatPopen.communicate() directly convert bytes to strings
Instead of only trying to convert to strings when stderr or stdout are
individually pulled out of communicate (e.g. communicate()[1]) just try
to convert both stderr and stdout into strings and return a new tuple.
This makes callsites that do `stdout, stderr = p.communicate()` a lot
simpler. Note that we try to convert to str independently and we don't
take the all-or-nothing approach like concat_streams does.