Fix scans of array views
When parallelizing scans, I accidentally broke scans of reindex and
slice expressions without realizing it.  This PR:

* fixes scans of slice expressions
* makes scans of reindex expressions continue to be serial for now by
  disabling forwarding of doiScan
* disables forwarding of doiScan on rank change expressions as well
  even though they wouldn't resolve since the array is multidimensional
  and we currently only support scans of 1D arrays

It also adds a test to lock in the current behavior when scanning
views and a test to demonstrate that scans of bools are currently
working (not necessarily related to this PR, but something I've
been meaning to do).