Massive rewrite of the eureka pattern demonstrator.
We read the 'found' indication way more than we write it, so use an
array of 'found' indicators, one per node, thus allowing completion
checks to be local.  Get rid of the DSIUtil internal calls to chunk
index ranges and do that in explicit code instead.  Don't create
daemon tasks for the per-CPU serial searches; they're unnecessary
and costly.