Increase the amount of oversubscription for cce whitebox testing
Increase from running 2 paratest instances to 6 for cce testing to try
and get testing to finish before the PE update window. cce testing runs
on kay-elogin, which is beefy, but is also a shared machine. Bump the
amount of oversubscription to try to finish before PE updates, but don't
use an insane amount since it's still a shared machine. paratest `nice`s
the start_test runs and we run with `CHPL_RT_OVERSUBSCRIBED=yes`, so
nodepara 6 shouldn't be too much.