Make it a fatal error if nightly fails to copy the temp log file
nightly uses a local log file stored in /tmp for speed and then copies
the results to some backed-up NFS location. If this copy fails then
we'll end up processing and reporting on last weeks results, which is
really confusing. Make it a fatal error if this copy fails. It failed
once last week, which is the first we know of, but this is something
that would have been hard to notice since it was silently ignored
before.