Get testVFlag passing with launcher=aprun.
testVFlag and testVerboseFlag didn't pass with launcher=aprun on the
Cray systems because that launcher produces an unexpected line of output
at the end, detailing resource usage for the job.  Here, adjust to this
by renaming the .goodstop file to .goodprog (it contains the program
output, so that's appropriate anyway) and concatenating the contents of
an optional .launcher-X.goodstop file to the end of the .good.  Augment
the .exec.out.tmp processing for launcher=aprun to remove the details
from the resource usage line, and finally, add .launcher-aprun.goodstop
files that match the resulting generic resource usage line.  These two
tests now pass with launcher-aprun.