Manually optimize for the 1-task case
Elliot found that my 'coforall tid in 1..numTasks' construct
was adding overhead relative to the original version of the
Chapel scan in the case where numTasks == 1 as in ISx where
we've already fully subscribed the machine.  While there are
long-term intentions to have the compiler automatically
optimize 1-task loops, since we don't have that yet, here
I'm manually optimizing it.