Add future to track issue with requiring a .c file
Require statements with .c files only work when the .c file being required lives
in the same directory as the compilation command is being executed from. When
the .c file lives in a different directory (even one specified via `-M`, or the
same directory as the .chpl file that relies on it), the Chapel compiler cannot
find it.