Add --no-warnings flags for these tests
While I was hoping that we could remove --no-warnings for all ISx
variants, it turns out that these forms still use a promotion (a
promoted read() on an atomic variable), so require it.  I was
tempted to add the warnings to the .good files, but since the
.goods are common across test variants, and the line numbers for
the warnings are not, that didn't work.  But at least we were
able to remove the directory-wide warning.