Add a test to lock this behavior in
This test explicitly sets CHPL_HWLOC=none in a normally =hwloc setting.  It then
attempts to override the setting via the compilation command, and link to the
generated library file (when the setting is back to =hwloc).  I verified that
this

1. did not allow the program to compile on master but does now, and
2. won't allow the program to compile if CHPL_HWLOC is left set to none (i.e.
if --hwloc=hwloc is not provided but the .compenv file is still present, the
program will fail)