Optimize stencil distribution's fast-follower
The forces elements being yielded from the stencil dist fast-follower to
be narrow, which significantly improves the generated code and results
in a 1% speedup for prk-stencil at 256 nodes.

Similar to #11644 -- prevent the compiler from widening elements being
yielded from the fast follower. This is similar to #11644, but the same
approach didn't work for PRK stencil. Previously, a slice of the local
array was taken, but the updateFluff() call in prk-stencil ends up
widening the slice routine so we still had wide pointers. Here we switch
from slicing to doing direct indexing in order to avoid that problem and
get narrow pointers.

This improves the generated code and eliminates wide pointers for the
`forall i in input do i += 1.0;` step in prk-stencil. It's not quite as
optimal as the reference version because we still have our `blk` offset
for 2D arrays and some other misc generated code, but it's a lot cleaner
than it was.