File a bug future showing that opaque arrays are lossy when resized
This test demonstrates that opaque arrays are currently lossy when
they're resized.  I believe that this is because their implementation
stores DefaultAssociative classes rather than a user-facing default
associative domain/array and that as a result, we're missing some of
the logic that implements resizing in the _array/_domain wrappers.