Fix problem with execution limiter in Python 3.7
This is a follow-on to #12597.
That PR added a check of an OSError's errno against os.errno.EEXIST.

But in Python 3.7, os.errno is not defined. Instead, error numbers
are moved to a top-level errno module.

This code change uses FileExistsError for Python 3 but it would also
be reasonable to look for the error numbers in the new place with
Python 3.7.