Make sure comm=ofi AMOs are coherent.
If a provider could support network AMOs for a given datatype only for
one or two of the non-fetching, fetching, and comparison atomic op
classes but not all three of those, the logic in comm=ofi would use
network AMOs for the supported op classes and drop back to Active
Messages and CPU AMOs for the unsupported op classes.  But that would
mix two kinds of initiators for a single datatype, potentially leading
to non-coherence.  Here, arrange to use network AMOs for all ops for a
given datatype iff the provider supports those for all three atomic op
classes, and otherwise use AMs and CPU AMOs for all ops.

(Note that we're not aware of any providers that have partial network
atomic op support as described here, but the new code is still the safer
way to go.  And the new logic is probably a little faster, too.)