Allow the unordered optimization to trigger for block distributed arrays
The compiler analysis special cases dsiAccess as a non-blocking call,
which we need to do because otherwise the locks used for the RAD opt
indicate synchronization. However, BlockDist has a helper nonLocalAccess
method that was thwarting the analysis. Add a pragma to be able to mark
functions that are safe for the optimization and apply it to
nonLocalAccess.

A better solution might be a pragma to indicate a critical section that
denotes internal synchronization only, but this is easy for now.