Update test of Replicated distributions to reflect proposed slicing semantics
With the proposed change in slicing semantics to have the slicing
domain govern the computation, this test changes in behavior since it
was slicing into a Block-distributed array using a DR domain.  Here,
I've created two versions of the slicing domain, one that's local /
DR, the other of which is created by slicing the distributed domain.
I then cloned a few of the interesting computations to use both the
local and distributed domain to capture the difference in behavior.

I think there's an open semantic question still about whether whether
the array or domain should "win" in terms of determining how a slice
executes, so this behavior could potentially change down the line as
we get more experience with the new behavior.  However, I think having
domains own slices is important enough to justify taking this step
in the meantime even if we adjust the behavior further going forward.