Update test to adjust to "slice by default rectangular" semantics
This test was relying on the fact that slicing a stencil-dist array
by a DR would create a stencil-dist domain which differs from most
domains in that it permits slightly OOB accesses.  However, with
the slicing semantic changes being undertaken in this PR, the domain
of that slice is now simply a DR which makes the OOB accesses
illegal.

As a result, I've updated this test to explicitly create stencil-dist
domains and use those to slice the array to retain the same behavior
as before while also retaining the original slice-by-DR cases to show
that they now result in OOB accesses.

As mentioned in my previous commit on this branch, I'm not 100% sure
that we won't go back to having slice-by-DR semantics end up being
more like the current behavior on master, but believe that this is
a worthwhile stepping stone to get into the world in which we use
existing domains when slicing rather than always copying them.