Separate functions that return nothing from void values
Add a new type: 'nothing' that is the type for functions that don't return
anything.  Update the compiler and modules to use it instead of 'void'
wherever it means a function that doesn't return anything. The names of the
compiler globals for this type are dtNothing and gNothing.

Changed the name of the void value from '_void' to 'voidValue' in the compiler
and modules.  The names of the compiler globals for this type are dtVoid and
dtVoidValue.