Ignore timeout files that are smaller than CHPL_TEST_TIMEOUT
Previously we would always use the timeout that was in a timeout file,
but this doesn't make a sense if we set CHPL_TEST_TIMEOUT to a value
larger than the one in the timeout file. Here we effectively switch
this to be `timeout = max(globalTimeout, timeoutFile)`, but we don't
want to do this for timeouts lower than the default (e.g. for cases
where we set a 30 second timeout so we fail fast.)