Move RandomStream locking/locking into helper routines
Move duplicated lock/unlock code into helper routines. Also make the locks void
when parSafe=false. An upcoming PR will prevent these sync locks from thwarting
RVF and it will be easier to do that when the logic is encapsulated. This also
makes it easier if we want to use atomic locks in the future or something.