Fix multilocale/syncsingle.prediff
This prediff was outdated (and not doing anything) as of #7343, which
compressed 0 entries out of comm diagnostics. The prediff effectively accepts
either 1 or 2 fast-ons from the last locale, but it was really fragile before
in terms of hard coding how many spaces and parenthesis were in the comm diags
output. This wasn't a problem for a while because this test seems stable with 2
fast-ons for most configurations, but gasnet-numa seems to vary between 1 and
2, so the prediff is needed again now that we're running this test under
gn-numa. There's probably a better way to do this (have the test self-verify
the comm like some other comm-count tests do), but this was pretty simple.