Move class subtyping check to canCoerce
from canDispatch. When passing a subclass to a superclass
argument, a cast is necessary in the generated C code.
One day this could possibly be removed with the LLVM backend,
but checking for it it in canDispatch just confuses matters today.

This commit moves that code to canCoerce and adjusts it to
handle nilable -> not nil, unmanaged -> borrow, etc.

While there, used a simpler style for some of the related methods.