Filter out extra messages for multilocale/local/diten/test_local2
This test has some nested on-stmts and the innermost one halts. We sporadically
get warnings like: "GASNET WARNING(Node 0): AMUDP_DrainNetwork(ep_t) returning
an error code: AM_ERR_RESOURCE" for this test. We used to get more errors for
any test that halted from a non-zero locale, but we don't see many of those
since the GASNet-EX upgrade. This test is only one that has seen been noisy and
that's probably from the nested on-stmt. Just add a prediff to filter out any
extra messages and only look for our expected error.