Fix the error messages for the multilocale library makefile
The ZMQ error previously triggered a different error due to the check being
defined before a build target.  This makes a build target to check and calls in
the `all` build, so it fires before building the executable or library.

Also, while here I noticed that the CHPL_MAKE_LAUNCHER error used the MLI
acronym - change that to "multilocale library".