Fix alwaysRVFpragma test
This test was having troubles when run under valgrind for --no-local
or multilocale runs.  I believe that these related more to the default
serialization routines being bit copy routines and messing up the
shared class references than because of anything broken in the
implementation of the always RVF pragma itself.  For that reason, I
rewrote the test to preserve the kind of locality checks I'd been
hoping to do but without writing serialization routines or using
classes.