Add AMOs to CommDiagnostics.
We've long wanted to include atomic operations in comm diagnostics.  Do
that here.  But param-disable printing atomic op counts by default,
because those counts are unstable due to our use of network atomics when
spin-waiting for parallel construct and on-stmt completion.

(Note that we would prefer to use a config const to control whether AMO
counts are printed, at execution time instead of compile time, but
initialization order constraints prevent using a config const here.
Basically, the problem is that the CommDiagnostics is initialized before
the String module is.)