Fix spawn-popen-wait to be more robust
This test is trying to check that a process openned with popen is "live"
- that is that its output arrives at the spawning process before the
process quits.

It previously implemented this check with sleeps.

This commit changes it to signal the spawned process (by creating a file)
to change its output. If the popen implementation ran the subprocess to
completion immediately (say), then this test would timeout because the
subprocess waits for the signal to complete.

Note that the shell script waiting for the file still includes a sleep
(since this is apparently the portable way to do that in a shell script).
Nonetheless the test should now be robust against various forms of system
interruption.