Get rid of some troublesome 'extern' declarators.
Remove the `extern` declarators on several runtime control variables.
The BSD-like 'common' semantics they create on Mac OS X causes trouble
if they don't have initializers, but adding otherwise superfluous
initializers just to solve that problem seems artificial.  So here,
simplify the whole mess: get rid of the 'extern' declarators, git rid of
the initializers, and get rid of the definitions in the .c files since
the declarations in the .h files create implied definitions anyway.

As a side effect, change the type of chpl_memTrack from chpl_bool to int
so that we can be absolutely assured that its initial value is "false"
in the C sense.