Enable the unordered compiler optimization by default
Turn on --optimize-forall-unordered-ops by default. This optimization
was added in 12269, but we left it off by default because the release
was so close.

I've looked at all the release/examples tests and all the multi-locale
benchmarks as a sanity check for when this optimization triggers and all
those cases are legal.