Throw --no-optimize-forall-unordered-ops for some bale variants
We were already tracking ordered vs explicit unordered vs compiler-opt
unordered for histogram and indexgather, so just keep the same reporting
by modifying compopts to maintain the old test behavior. This is simpler
than RA since this is a study version instead of in the release dir so I
don't mind the "default" version not representing the default behavior.