Add a test that creates more remote tasks than there is memory for
Add a test that creates thousands of remote tasks (more task stacks than
there is physical memory for if they actually ran concurrently.)  I
believe this is a test we want to work under qthreads since there is no
user-induced task yields. It currently fails under both ugni and gasnet
because of task yielding in the comm layers.