Avoid comm/compute overlap in ugni for "minimal-comm" tasks
Avoid comm/compute overlap for FMA in ungi for tasks that haven't issued at
least 100 FMA. This is a little bit odd, but it prevents us from doing
comm/compute overlap for tasks that haven't done much comm. This allow us to
run an arbitrary number of tasks with no user-induced yields, but still get
comm/compute overlap for RA.

Here we use task-local-storage to track the amount of comm. Accessing task-local
storage is pretty cheap (~ half the cost of a processor atomic) so this adds
virtually no overhead.