Execute defer actions in an iterator upon break from its loop
Improves #6540, #7897. Reverts #12840.

Upon breaking out from a for-loop, call back into the loop's iterator
to perform the defer actions needed for its current `yield`.

Breaking out from a for-loop happens upon executing a `break`, `return`,
or `throw` in the loop body.

The "defer actions" for a given `yield` in an iterator are the actions
to be executed upon exiting the iterator immediately after executing
the `yield`. This is what would happen if we placed a `return` there.

For example, in this code:

```chpl
for idx in myIter() {
  if ... then
    break;
}

iter myIter() {
  defer { myCleanup(); }
  for idx2 in myIter2() {
    var r = new MyRecord();
    yield idx2;
    // the defer actions for the above 'yield' are the ones we'd execute
    // if we placed a `return` here
  }
}
```

when the for-loop executed `break`, we used to "forget" about `myIter`
immediately. However, the code in `myIter` requires certain cleanup
after executing the yield:

* deinitialize `r` ,
* call `myCleanup()` ,
* deinit the _iteratorClass for `myIter2()`, when applicable.

These are now properly executed, whether `myIter()` is inlined or not.

While there:

* Restructure some existing code in lowerIterators for improved readability.

* view.cpp/debugging instrumentation
  - improve how `list_view` and `nprint_view` display a `GotoStmt`
  - tweak the output of `fnsWithName` and `whocalls`
  - add gotoTagToString()

#### A war story

When I thought I was finished, I noticed that sungeun/fannkuch-redux.chpl
was leaking 100MB of iterator classes. Oops! Turned out that IC.more
remains at `1` for any not-inlined single-loop single-yield iterator.
Therefore when a `break` is executed, negating `IC.more` did not lead
to executing an IBB within `advance`. Instead, `advance` initialized the IC
one more time. This is why I added "this.more=2;" to buildZip2().

So, surprisingly, a memory leak was actually an indication of
a correctness issue.

#### Future work

* Handle recursive iterators.

* Ensure the defer actions are executed upon a `throw` in the for loop.

#### Mem leak improvements

* no leaks:

arrays/deitz/jacobi-no-inline.chpl
associative/bharshbarg/arrays/diffDomZip.chpl
memory/figueroa/LeakedMemory6.chpl
types/records/ferguson/deinit/break-in-for-iter1
types/records/ferguson/deinit/break-in-for-iter2

* now leaks 1 record where before it leaked 4 records:

types/records/ferguson/leak-futures/iterate-loop-break2

* reverted #12840 without re-introducing the leaks:

library/packages/VectorList/test-pariter.chpl