Cast to void* to avoid GCC 9 warning
We see warnings with GCC 9 compilers from -Werror=class-memaccess
due to memcpy/memset on a class instance pointer. This is OK for
the types we do it on (since they have no virtual dispatch tables
and are actually POD in practice) so use a cast to quite the warning.

See also https://gcc.gnu.org/onlinedocs/gcc/C_002b_002b-Dialect-Options.html
which indicates that casting to void* can quiet the warning.