Use strcpy/strcat instead of strncpy/strncat with src length
GCC 9 gives a warning for this code because the length supplied for
strncpy/strncat is the *source* buffer length rather than the amount of
bytes remaining in the destination buffer.

Using strncpy/strncat in this way is actually just the same as using
strcpy/strcat. So, this commit changes the code to use strcpy/strcat.
This removes the warning under GCC 9.