Base XC hugepage heap management on Aries network, not gni provider.
We have been deciding based on the provider whether to build a fixed
heap on hugepages.  With the upcoming switch to capability-centric
fabric selection rather than provider-centric, base this on the network
architecture instead.  (It's still actually the provider than needs it,
but at the time we have to make the decision we don't yet know what the
provider will be.  This is really just a sanity configuration anyway,
though, so that's okay.)