Choose fabrics based on desired capabilities, not provider name.
We were slightly misusing the libfabric interface: I think it's expected
that users should ask for the capabilities they want (in the hints
passed to fi_getinfo()) and then perhaps filter on the provider, rather
than forcing the provider up front.  Here, stop including any provider
in the hints unless the user specifies one.  However, do adjust the
hints based on what we surmise the network to be, because it still seems
as though giving the additional information can produce a better result.

Related to this, for now skip tcp providers because they seem to be
exhibiting a behavior we haven't seen before.  We didn't use to (try to)
use them because we forced sockets providers as a default and never
tested specifying tcp ones, so we're not losing anything here yet.  This
does need fixing eventually, though, because the tcp providers are
nowadays expected to be the best-performing non-proprietary ones.  But
it's not necessary for the current work.

While here, add debug printing of the hints that are input to the
fi_getinfo() call.

While here, fix a spot where we could end up trying to use a completion
counter associated with a TX context that actually has a CQ instead.