Add CHPL_COMM_USE_GDB support to many launchers.
The amudprun launcher for comm=gasnet has long had the ability to launch
each Chapel program instance as the subject of a debugger running in an
xterm, rather than launching each program instance directly.  Here, move
that support so that a significant fraction of all the launchers can use
it, namely: amudprun, aprun, gasnetrun_ibv, gasnetrun_mpi, pbs-aprun,
mpirun4ofi, slurm-srun, and smp.

Big caveat: it is the user's responsibility to make sure things are set
up so the xterm which is run in the target environment can open its X11
window in the launch environment.

(I started out just wanting to add this capability in the mpirun4ofi
launcher for my own benefit, but then realized that it wouldn't be too
much harder to add it to common code that affected many more launchers.)