Fix some futures/.bad mismatches about tuple expansion
There is now an error about bad tuple expansions instead of letting it
pass through in an undefined way.

One of these .futures was asking for such an error message, so unfuturized it.

The other two want tuple expansion inside of parenthesis to result in a tuple,
but they now get the bad tuple expansion error. Update their .bad files. These
expect `((...tup))` to be a tuple, but it's currently just a parenthesized
expanded tuple.  To get a tuple, you can add a comma: `((...tup),)`.  I didn't
add the commas because it slightly changes the intent of the tests.