Clean this up a bit
This is pretty weird:  If we try to resolve a block-distributed
array, chpl__serialize() returns a 2*int(64) which should only
happen for slices.  So maybe it's a slice of a block-distributed
array?  But then it can't deserialize() using that same value...
Could have something to do with static vs. dynamic types?