Improve the speed of parallel safe random streams
Previously, we used a sync lock but that's extremely slow, especially in
serial/uncontested code. Transition to using a test-and-testAndSet lock.
When contested a testAndSet lock is slower than a sync lock because of
cache contention but a test-and-testAndSet lock is faster (#13012)

For the performance test added in #13009, this results in a 17x speedup for
streams accessed serially and a 7x speedup for streams accessed concurrently.

| config               | before | after  |
| -------------------- | ------ | ------ |
| serial (not parSafe) |  0.06s |  0.06s |
| serial (parSafe)     |  8.50s |  0.50s |
| parallel (parSafe)   | 34.50s |  4.50s |

Note that when parallel safety is not needed, using `parSafe=false` is
still 10x faster, but this significantly reduces the default overhead.