Add parallel one locale amo test
The tests the performance of all tasks on one locale updating an atomic on
another locale with the crossproduct of static/dynamic and ordered/unordered.
This is meant to help with performance tuning in a simpler benchmark and also
showcase that network ops from numa domain 1 are slower than those from numa
domain 0 as indirectly shown by dynamic scheduling being faster (static
partition of data isn't the best)

This test can also be run in other modes to show oversubscription benefits, but
that's not a config run nightly.