Add .futures about extern functions that return void
Add .futures about extern functions that take array arguments and explicitly
return 'void' through type aliases.

We create a wrapper function to turn the array argument into a simple pointer.
The wrapper returns the result of calling the extern function, but if it
returns 'void' we shouldn't return it. If the return type is a type alias,
we don't currently have a good way to detect that it is 'void' at the time
the wrapper is created.