Add some error messages for the new hidden flag
- If we see it twice, that's because the user added it themselves, so we tell
them it is for internal use only
- If we only see the flag and it is the last argument (not followed by the
connection information), tell them it needs to be followed by connection
information
- If we see the flag and what follows it does not start with `tcp://`, then it
is likely being used improperly (either by not having connection information
follow it, or by having a bad string follow it), so tell the user not to.

In any case, only the last use of the flag would matter in a multilocale library
run (other uses get dropped on the floor) and we ensure that we provide the last
use of the flag when we look for it, so it shouldn't pose a problem.