Clean up Makefile output a bit...
For awhile, our Makefiles have been inconsistent about whether to echo
the recursive `cd ... && make ...` commands or not, and the output for
a clean build has also gotten a bit messier.  This PR tries to clean
things up a bit by adding high-level echo commands to say what we're
building (and sometimes why) and squashing the make commands
themselves.  The output for a clean GASNet + LLVM build after this,
for example, looks like:

```
Making the compiler...
Making LLVM...
make[3]: Nothing to be done for `llvm'.
***** adt/ *****
***** AST/ *****
***** backend/ *****
***** codegen/ *****
***** ifa/ *****
***** llvm/ *****
***** main/ *****
***** optimizations/ *****
***** parser/ *****
***** passes/ *****
***** resolution/ *****
***** util/ *****
***** ./ *****
Speculatively attempting to build re2
make[3]: Nothing to be done for `re2'.
Speculatively attempting to build re2
make[3]: Nothing to be done for `re2'.
make[2]: Nothing to be done for `try-gmp'.
make[2]: Nothing to be done for `try-gmp'.
Making the runtime...
Making hwloc...
make[3]: Nothing to be done for `hwloc'.
Making jemalloc...
make[3]: Nothing to be done for `jemalloc'.
Making third-party-packages...
make[3]: Nothing to be done for `re2'.
make[3]: Nothing to be done for `qthread'.
make[3]: Nothing to be done for `hwloc'.
make[3]: Nothing to be done for `jemalloc'.
make[3]: Nothing to be done for `gmp'.
***** src/comm/ *****
***** src/comm/none/ *****
***** src/mem/ *****
***** src/mem/jemalloc/ *****
***** src/tasks/ *****
***** src/tasks/qthreads/ *****
***** src/threads/ *****
***** src/threads/none/ *****
***** src/timers/ *****
***** src/timers/generic/ *****
***** src/topo/ *****
***** src/topo/hwloc/ *****
***** src/qio/regexp/re2/ *****
***** src/qio/auxFilesys/hdfs/ *****
***** src/qio/auxFilesys/curl/ *****
***** src/qio/ *****
***** src/ *****
***** ./ *****
Making the runtime for LLVM...
Making hwloc...
make[3]: Nothing to be done for `hwloc'.
Making jemalloc...
make[3]: Nothing to be done for `jemalloc'.
Making third-party-packages...
make[3]: Nothing to be done for `qthread'.
make[3]: Nothing to be done for `re2'.
make[3]: Nothing to be done for `gmp'.
make[3]: Nothing to be done for `jemalloc'.
make[3]: Nothing to be done for `hwloc'.
***** src/comm/ *****
***** src/comm/none/ *****
***** src/mem/ *****
***** src/mem/jemalloc/ *****
***** src/tasks/ *****
***** src/tasks/qthreads/ *****
***** src/threads/ *****
***** src/threads/none/ *****
***** src/timers/ *****
***** src/timers/generic/ *****
***** src/topo/ *****
***** src/topo/hwloc/ *****
***** src/qio/regexp/re2/ *****
***** src/qio/auxFilesys/hdfs/ *****
***** src/qio/auxFilesys/curl/ *****
***** src/qio/ *****
***** src/ *****
***** ./ *****
Making the modules...
Updating TAGS...
Making the modules for LLVM...
Updating TAGS...
```