Fix broken fi_cntr usage.
I had misread the documentation regarding libfabric completion counters.
I mistook fi_cntr_read() as returning the number of completions counted
since the last time that function was called on the given endpoint or
context.  It actually just returns the value of the counter, which
starts at 0 when it is created and increments each time a completion
occurs, for the life of the program.  Fix this.

This fixes the "Too many callers" error check failure in the commDiags
test with comm=ofi.  There are still failures due to variation in the
output for that test, though, which occur when locale 0 is the last one
to arrive at the barrier.  In 1000 runs on our IB cluster that happened
2 times.  I'll fix that separately.  (This also afflicts comm=ugni, and
is a test problem rather than a problem in either comm layer.)

While here, I made some other improvements, moving the wait-for-counter
code into a common function and making that and the wait-for-CQ function
fairly similar, passing the transmit context info from the AM handler
into the AM AMO handler to save the latter the trouble of "allocating"
it (which in this circumstance is a slow no-op), improving commonality
in the debug output, and removing some unused counters.

As a beneficial side effect, this also gets rid of most of the timeouts
in slurm-based comm=ofi full testing on our IB cluster.  There were only
2 when I tested this PR, vs. 16-20 or so in runs over the last several
weeks.